qemuDomainUndefineFlags: Forbid undefine of active domain with NVRAM
https://bugzilla.redhat.com/show_bug.cgi?id=1467245

Currently, there's a bug when undefining a domain with NVRAM
store. Basically, the unlink() of the NVRAM store file happens
during the undefine procedure iff domain is inactive. So, if
domain is running and undefine is called the file is left behind.
It won't be removed in the domain cleanup process either
(qemuProcessStop). To avoid this forbid undefining domain with
NVRAM file.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <995ee582b9b42d2548a4c3c8e6067e5b3a438b38.1502279713.git.mprivozn@redhat.com>