util: fix Coverity complaint (and actual bug) in virHostdevReadNetConfig()
Commit 9a94af6d restructured virHostdevReadNetConfig() so that it
would manually set ret = 0 after successfully reading the device's
config, but the "ret = 0" was erroneously placed outside of an "else"
clause, meaning that the the value of ret set in the "if" clause was
unnecessarily and incorrectly overwritten.

This patch moves ret = 0 into the else clause, which should silence
Coverity.
Message-Id: <20170813154431.27234-2-laine@laine.org>
Reviewed-by: John Ferlan <jferlan@redhat.com>