util: Introduce virObjectLookupKeys
Add a new virObjectLockable child virObjectLookupKeys that can be
used by various driver/vir*obj consumers as the means to define the
lookup keys for an object.

The API requires that at least the @key1 argument is provided as
non-NULL thus ensuring that there's at least one key. The keys will
be used as input to a soon to be introduced hash table object that
will support one or two hash tables and manage the various add,
search, and remove functionality using the keys.

The initial consumer will be the driver/vir*obj.c APIs which will
make use of the keys via the @UUID or @Name that is used to uniquely
describe the object.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20170818215041.8118-3-jferlan@redhat.com>