util: Introduce virObjectLookupHashForEach
Introduce an API to use the virHashForEach API to go through
each element of the LookupHash calling a callback with a LookupKey
object entry from the LookupHash in order for it to be processed.

Create a common structure _virObjectLookupHashForEachData to define
the various pieces of data needed by the callback consumers. Fields
described in the API function description section.

Upon successful completion, the data.nElems is returned to the
caller and possibly various fields within the structure filled
in depending on the callers need as follows:

  vir{Object}NumOf{Elem}
  vir{Object}Get{Key}

      => Callback function can be combined since generally all that
         the NumOf cares about is counts, but the Get function also
         adds some {Key} to a returned list of @names.

  vir{Object}Export
      => When @maxElems == -1, will cause allocation into data->elems
         of an array sized by the hash table size which will be used
         by the callback function to store addresses of objects specific
         to each {Object} that are typically associated with the
         virConnectListAll* API's from the virGet{Object}() calls.
         The @flags argument is typically used for additional filtering
         via vir{Object}Match or vir{Object}Filter APIs found in various
         vir*obj modules. Upon return from the call, the calling function
         must then move the data->elems into the returned structure.

It is up to the callback function to set @error in the event of an error
during processing resulting in calling the VIR_FREE() for each of the
@nElems already in the array.

When an error occurs, any collection done by the callback functions into
the @elems array is discarded.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20170818215041.8118-8-jferlan@redhat.com>