util: Introduce virObjectLookupHashClone
A convenience API that will utilize the virHashForEach API for the
LookupHash in order to create a clone/copy. Primary consumer is the
interface driver which has a desire to save off a copy of its only
hash table in order to possible restore it if something goes wrong
during processing.

The callback function's primary purpose is to copy anything within
the local LookupKeys into the target.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20170818215041.8118-12-jferlan@redhat.com>