util: Introduce virObjectLookupHashPrune
A convenience API that will utilize the virHashForEach API for the
LookupHash in order to remove elements from the hash table(s) that
match some requirement from the callback.

NB: Once elements are removed from objsKey1 - if something goes wrong
in objsKey2, the best that can be done is to issue a warning and hope
for the best.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20170818215041.8118-16-jferlan@redhat.com>