network: Use virObjectLookup{Keys|Hash}
Use the virObjectLookupKeys in _virNetworkObj and use the
virObjectLookupHash in _virNetworkObjList.

Convert the code to use the LookupHash object and APIs rather
than the local code and usage of virHash* calls.

Since the _virNetworkObjList only contains the @parent object
the virClassNew must be removed from OnceInit because instantiation
would fail since the object size would be the same as the parent
object size.

Usage of HashLookup{Find|Search} API's returns a locked/reffed
object so need to remove virObjectLock after FindBy*Locked calls.

Use the def->name as a second key to the LookupHash to make for
faster lookup's by name (instead of using Search on uuid key
and matching the name.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20170818215041.8118-18-jferlan@redhat.com>