util: Introduce virObjectLookupHash
Define an object meant to combine/manage the aspects of the various
_vir*ObjList structs into one common purpose list management system
that will store and manage _vir*Obj virObjectLockable objects rather
than having multiple structs and API's to do that.

The object will use the virObjectRWLockable class as the base and may
have one or two hash tables that will be keyed by UUID or Name based
on the needs of the consumer stored in:

    objsUUID -> Hash table for storing objects for lookup by UUID
    objsName -> Hash table for storing objects for lookup by Name

The virObjectLookupHashNew will require the consumer to provide which
hash tables are to be used via virObjectLookupHashNewFlags.

A secondary benefit of self locking hash tables is each driver then
does not have to keep a higher level driver lock for interactions with
the object storage since the object itself can manage locking as needed.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20170823212211.4693-3-jferlan@redhat.com>