util: Introduce virObjectLookupHash{Add|Remove}
Add a pair of API's to add/remove an object (virObjectLockable)
to/from the LookupHash object. The caller must check return status
and handle failure properly for the Add. The Remove API callers are
all void functions, so only report the problem and ignore the attempt
to remove if for some reason the passed @tableobj is not as expected.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20170823212211.4693-4-jferlan@redhat.com>