util: Introduce virObjectLookupHashFind[Locked]
These API's will use the virHashLookup in order to find the object in
the lookup hash object by the specified @key. If two hash tables exist
in the hash table object, then both will be searched for the key. Both
API's will call an virObjectLookupHashFindInternal in order in handle
the fetch.

The virObjectLookupHashFindLocked is the primary workhorse and should
only be called externally if the caller has taken the proper RW LookupHash
lock. This is necessary in some paths, such as during Add/Assign processing
where getting the RW Write lock early is required to ensure no other thread
attempts to create/add the same object

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20170823212211.4693-5-jferlan@redhat.com>