util: Introduce virObjectLookupHashSearch{UUID|Name}[Locked]
Add common object API wrappers to use the virHashSearch API to
search the LookupHash for specific data defined in the @opaque
parameter. Once data is found, the search would end and the
object that is represented is returned locked with it's reference
's reference count incremented.

The virObjectLookupHashSearch{UUID|Name}Locked is the workhorse, but
similar to the virObjectLookupFind* APIs may be required an Add or
AssignDef caller has the Write lock already to ensure no other thread
will grab the lock and add the same or competing object.

It is up to the caller to unlock and lower the refcnt once done
with the object and of course handle a NULL return indicating
no object found.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20170823212211.4693-7-jferlan@redhat.com>