nodedev: Use virObjectLookupHash
Use the virObjectLookupHash in _virNodeDeviceObjList. Convert
the code to use the LookupHash object and APIs rather than
code within this module that uses virHash* calls.

Since the _virNodeDeviceObjList only now contains the @parent
object, the virClassNew must be removed from OnceInit because
instantiation would fail since the object size would be the same
as the parent object size.

Usage of HashLookup{Find|Search} APIs returns a locked/reffed
object so remove the virObjectLock after FindBy*Locked calls.

The only function requiring taking a lock is the Add function
since it needs to be synchronized in such a way to avoid
multiple threads attempting to add the same named node device
at the same time.

The NumOfDevicesCallback and GetNamesCallback can use the same
callback function with the only difference being the filling in
of the @names array from the passed @data structure if it exists.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20170823212211.4693-8-jferlan@redhat.com>