util: Introduce virObjectLookupHashClone
A convenience API that will utilize the virHashForEach API for the
LookupHash in order to create a clone/copy. Primary consumer is the
interface driver which has a desire to save off a copy of its only
hash table in order to possible restore it if something goes wrong
during processing.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20170823212211.4693-10-jferlan@redhat.com>