interface: Use virObjectLookupHash
Use the virObjectLookupHash in _virInterfaceObjList. Convert the code
to use the LookupHash object and APIs rather than the local forward
linked list processing.

Usage of HashLookup{Find|Search} API's is via a callback mechanism
and returns a locked/reffed object.

The Clone API will make use of the ForEach functionality in
copying whatever is in one LookupHash list into the destination.

The only function requiring taking a lock is the AssignDef
function since it needs to be synchronized in such a way to
avoid multiple threads attempting to add the same named
object at the same time.

The NumOfInterfaces and GetNames can use the same callback
function with the only difference being the filling in of the
@names array from the passed @data structure if it exists.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20170823212211.4693-12-jferlan@redhat.com>