util: Introduce virObjectLookupHashPrune
A convenience API that will utilize the virHashForEach API for the
LookupHash in order to remove elements from the hash table(s) that
match some requirement from the callback.

NB: Once elements are removed from objsUUID - if something goes wrong
in objsName, the best that can be done is to issue an error message
and hope for the best.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20170823212211.4693-14-jferlan@redhat.com>