util: Introduce virFileWaitForAccess
Since we have a number of places where we workaround timing issues with
devices, attributes (files in general) not being available at the time
of processing them by calling usleep in a loop for a fixed number of
tries, we could as well have a utility function that would do that.
Therefore we won't have to duplicate this ugly workaround even more.

Signed-off-by: Erik Skultety <eskultet@redhat.com>
Message-Id: <99af32dc75bfc0891e4e06dbf9cc59f5395e4bdc.1503573079.git.eskultet@redhat.com>