qemu: Handle host devices not being available better
We can't retrieve the isolation group of a device that's
not present in the system. However, it's very common for
VFs to be created late in the boot, so they might not be
present yet when libvirtd starts, which would cause the
guests using them to disappear.

Moreover, for other architectures and even ppc64 before
isolation groups were introduced, it's considered
perfectly fine to configure a guest to use a device
that's not yet (or no longer) available to the host,
with the obvious caveat that such a guest won't be able
to start before the device is available.

In order to be consistent, when a device's isolation
group can't be determined fall back to not isolating it
rather than erroring out or, worse, making the guest
disappear.

Resolves: https://bugzilla.redhat.com/show_bug.cgi?id=1484254

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Message-Id: <20170825161537.8331-1-abologna@redhat.com>