qemu: Introduce qemuDomainPrepareDiskSource
Introduce a function to setup any TLS needs for a disk source.

If there's a configuration or other error setting up the disk source
for TLS, then cause the domain startup to fail.

For VxHS, if cfg->haveTLS is set, then TLS creds will be added
automatically to every VxHS disk that didn't specify "tls='no'"
in the domain XML. Additionally, if the domain XML has "tls='yes'",
but cfg->haveTLS is not set, then issue a configuration error.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20170830224613.24738-11-jferlan@redhat.com>